lvs_vip: bash script invoked by keepalived for master/slave transitions
Invoked by keepalived from master/slave to slave/master transition to
adapt network configuration when load balancer act as real server too.

Add VIP address on localhost for slave to accept incoming packets sent
to the VIP address.

Drop arp requests for the VIP in order to avoid slaves to acquire the
VIP and alter arp responses to force real IP address instead of VIP.

Add an iptables rule on the master to mark incoming packets on the VIP
address to be handled by LVS.